# docs/handoff/auth-2fa.openapi.yaml
openapi: 3.0.3
info:
  title: InventoryPro — Contrato de Autenticación con 2FA (estado objetivo)
  version: "1.1.0"
  description: >
    Contrato del flujo de login tras remediar el hallazgo H-02 (2FA decorativo). Fuente de
    verdad para la tarea P0-2. El login pasa a ser de dos pasos cuando el usuario tiene 2FA
    activo: el password por sí solo ya NO concede el JWT de sesión.
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Paso 1 — Autenticar con email y contraseña
      description: >
        Verifica credenciales. Si el usuario NO tiene 2FA habilitado, devuelve directamente el
        access_token (comportamiento previo, sin cambios). Si el usuario TIENE 2FA habilitado,
        NO emite el JWT de sesión: devuelve `requires2FA: true` y un `twoFactorToken` efímero
        que sólo sirve para el paso 2.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Credenciales válidas. Una de dos formas según el estado de 2FA.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LoginSuccess'
                  - $ref: '#/components/schemas/LoginRequires2FA'
              examples:
                sin2FA:
                  summary: Usuario sin 2FA
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    user:
                      id: "uuid"
                      email: "admin@inventorypro.com"
                      firstName: "Admin"
                      lastName: "Principal"
                      role: "ADMIN"
                con2FA:
                  summary: Usuario con 2FA habilitado
                  value:
                    requires2FA: true
                    twoFactorToken: "eyJ...token-efimero-pre-2fa..."
        '401':
          description: Credenciales inválidas o usuario desactivado (mensaje genérico).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/2fa/login:
    post:
      tags: [Auth]
      summary: Paso 2 — Validar el código TOTP y emitir el JWT de sesión
      description: >
        Recibe el twoFactorToken del paso 1 y el código TOTP de 6 dígitos. Si el código es
        válido, emite el access_token definitivo. El twoFactorToken es de un solo uso / vida
        corta y NO da acceso a endpoints protegidos por sí mismo.
      operationId: loginTwoFactor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorLoginDto'
      responses:
        '200':
          description: Código válido. Devuelve el access_token de sesión.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccess'
        '400':
          description: Código TOTP inválido (mensaje genérico).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: twoFactorToken inválido o expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/2fa/generate:
    post:
      tags: [Auth]
      summary: Generar secreto TOTP y QR (requiere sesión)
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: Secreto y QR para escanear en la app autenticadora. }
        '401': { description: No autenticado. }
  /auth/2fa/enable:
    post:
      tags: [Auth]
      summary: Habilitar 2FA tras verificar un código (requiere sesión)
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, example: "123456" }
      responses:
        '200': { description: 2FA habilitado. }
        '400': { description: Código inválido. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginDto:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    TwoFactorLoginDto:
      type: object
      required: [twoFactorToken, code]
      properties:
        twoFactorToken:
          type: string
          description: Token efímero devuelto por /auth/login en el paso 1.
        code:
          type: string
          pattern: '^[0-9]{6}$'
          example: "123456"
    LoginSuccess:
      type: object
      required: [access_token, user]
      properties:
        access_token: { type: string }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string }
            firstName: { type: string }
            lastName: { type: string }
            role: { type: string, enum: [ADMIN, STAFF] }
    LoginRequires2FA:
      type: object
      required: [requires2FA, twoFactorToken]
      properties:
        requires2FA:
          type: boolean
          enum: [true]
        twoFactorToken:
          type: string
          description: Token efímero (≈5 min) para el paso 2. No es el JWT de sesión.
    ErrorResponse:
      type: object
      properties:
        statusCode: { type: integer }
        message: { type: string }
        error: { type: string }